Properly purge state groups tables when purging a room
