+++ changelog.d/18024.bugfix
@@ -1,0 +1,1 @@
+Properly purge state groups tables when purging a room with the admin API.
